Added VBProjectile and VBProjectileShooter as the base for reusable weapon scenes
This commit includes the functionality for a more general projectile and projectile shooter that can be initialized from the outside.
These systems aren't ready due to poorly thought-out coupling between these types for initialization of these types. Warnings are thrown until this is fixed.

Also updated collision layers for the test player and target dummy.

Also a lot of documentation updates